model 2 models from Mining
