+++ src/main/java/me/joshmcfarlin/cryptocompareapi/api/Mining.java
@@ -3,6 +3,8 @@ package me.joshmcfarlin.cryptocompareapi.api;
 import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
 import me.joshmcfarlin.cryptocompareapi.models.mining.Coin;
+import me.joshmcfarlin.cryptocompareapi.models.mining.Contract;
+import me.joshmcfarlin.cryptocompareapi.models.mining.Contracts;
 import me.joshmcfarlin.cryptocompareapi.models.mining.Equipment;
 import me.joshmcfarlin.cryptocompareapi.utils.CallTypes;
 import me.joshmcfarlin.cryptocompareapi.utils.Connection;
@@ -40,330 +42,5 @@ public class Mining {
         return new Gson().fromJson(r, Equipment.class);
     }
 
-    /**
-     * Represents mining contracts
-     */
-    public class Contracts {
-        /**
-         * The response provided by the API
-         */
-        @SerializedName("Response")
-        private String response;
 
-        /**
-         * Message provided by the API
-         */
-        @SerializedName("Message")
-        private String message;
-
-        /**
-         * Mining data
-         */
-        @SerializedName("MiningData")
-        private Map<Integer, Contract> contracts;
-
-        /**
-         * Coin data
-         */
-        @SerializedName("CoinData")
-        private Map<String, Coin> coins;
-
-        /**
-         * Response type
-         */
-        @SerializedName("Type")
-        private int type;
-
-        /**
-         * {@link Contracts#response}
-         */
-        public String getResponse() {
-            return response;
-        }
-
-        /**
-         * {@link Contracts#message}
-         */
-        public String getMessage() {
-            return message;
-        }
-
-        /**
-         * {@link Contracts#contracts}
-         */
-        public Map<Integer, Contract> getContracts() {
-            return contracts;
-        }
-
-        /**
-         * {@link Contracts#coins}
-         */
-        public Map<String, Coin> getCoins() {
-            return coins;
-        }
-
-        /**
-         * {@link Contracts#type}
-         */
-        public int getType() {
-            return type;
-        }
-
-        /**
-         * Represents a mining contract
-         */
-        public class Contract {
-            /**
-             * ID of the equipment
-             */
-            @SerializedName("Id")
-            private int id;
-
-            /**
-             * Parent ID of the equipment
-             */
-            @SerializedName("ParentId")
-            private int parentID;
-
-            /**
-             * Equipment company
-             */
-            @SerializedName("Company")
-            private String company;
-
-            /**
-             * Url
-             */
-            @SerializedName("Url")
-            private String url;
-
-            /**
-             * Url of the logo
-             */
-            @SerializedName("LogoUrl")
-            private String logoURL;
-
-            /**
-             * Equipment name
-             */
-            @SerializedName("Name")
-            private String name;
-
-            /**
-             * Represents if recommended equipment
-             */
-            @SerializedName("Recommended")
-            private boolean recommended;
-
-            /**
-             * Represents if sponsored equipment
-             */
-            @SerializedName("Sponsored")
-            private boolean sponsored;
-
-            /**
-             * Affiliate url
-             */
-            @SerializedName("AffiliateURL")
-            private String affiliateURL;
-
-            /**
-             * Equipment algorithm
-             */
-            @SerializedName("Algorithm")
-            private String algorithm;
-
-            /**
-             * Number of hashes per second the equipment can reach
-             */
-            @SerializedName("HashesPerSecond")
-            private String hashesPerSecond;
-
-            /**
-             * Cost of the equipment
-             */
-            @SerializedName("Cost")
-            private double cost;
-
-            /**
-             * Currency of the cost
-             */
-            @SerializedName("Currency")
-            private String currency;
-
-            /**
-             * Cost of the fee
-             */
-            @SerializedName("FeeValue")
-            private double feeValue;
-
-            /**
-             * Currency of the fee
-             */
-            @SerializedName("FeeValueCurrency")
-            private String feeValueCurrency;
-
-            /**
-             * Length of the contract
-             */
-            @SerializedName("ContractLength")
-            private String contractLength;
-
-            /**
-             * Currencies available to be mined with the equipment
-             */
-            @SerializedName("CurrenciesAvailable")
-            private String currenciesAvailable;
-
-            /**
-             * Logos of the currencies available for mining
-             */
-            @SerializedName("CurrenciesAvailableLogo")
-            private String currenciesAvailableLogo;
-
-            /**
-             * Currencies available to be mined with the equipment
-             */
-            @SerializedName("CurrenciesAvailableName")
-            private String currenciesAvailableName;
-
-            @Override
-            public String toString() {
-                return String.format("%s - %s", company, name);
-            }
-
-            /**
-             * {@link Contract#id}
-             */
-            public int getId() {
-                return id;
-            }
-
-            /**
-             * {@link Contract#parentID}
-             */
-            public int getParentID() {
-                return parentID;
-            }
-
-            /**
-             * {@link Contract#company}
-             */
-            public String getCompany() {
-                return company;
-            }
-
-            /**
-             * {@link Contract#url}
-             */
-            public String getUrl() {
-                return url;
-            }
-
-            /**
-             * {@link Contract#logoURL}
-             */
-            public String getLogoURL() {
-                return logoURL;
-            }
-
-            /**
-             * {@link Contract#name}
-             */
-            public String getName() {
-                return name;
-            }
-
-            /**
-             * {@link Contract#recommended}
-             */
-            public boolean isRecommended() {
-                return recommended;
-            }
-
-            /**
-             * {@link Contract#sponsored}
-             */
-            public boolean isSponsored() {
-                return sponsored;
-            }
-
-            /**
-             * {@link Contract#affiliateURL}
-             */
-            public String getAffiliateURL() {
-                return affiliateURL;
-            }
-
-            /**
-             * {@link Contract#algorithm}
-             */
-            public String getAlgorithm() {
-                return algorithm;
-            }
-
-            /**
-             * {@link Contract#hashesPerSecond}
-             */
-            public String getHashesPerSecond() {
-                return hashesPerSecond;
-            }
-
-            /**
-             * {@link Contract#cost}
-             */
-            public double getCost() {
-                return cost;
-            }
-
-            /**
-             * {@link Contract#currency}
-             */
-            public String getCurrency() {
-                return currency;
-            }
-
-            /**
-             * {@link Contract#feeValue}
-             */
-            public double getFeeValue() {
-                return feeValue;
-            }
-
-            /**
-             * {@link Contract#feeValueCurrency}
-             */
-            public String getFeeValueCurrency() {
-                return feeValueCurrency;
-            }
-
-            /**
-             * {@link Contract#contractLength}
-             */
-            public String getContractLength() {
-                return contractLength;
-            }
-
-            /**
-             * {@link Contract#currenciesAvailable}
-             */
-            public String getCurrenciesAvailable() {
-                return currenciesAvailable;
-            }
-
-            /**
-             * {@link Contract#currenciesAvailableLogo}
-             */
-            public String getCurrenciesAvailableLogo() {
-                return currenciesAvailableLogo;
-            }
-
-            /**
-             * {@link Contract#currenciesAvailableName}
-             */
-            public String getCurrenciesAvailableName() {
-                return currenciesAvailableName;
-            }
-        }
-    }
 }
